+++ bsd/memory_openbsd.cc
@@ -26,7 +26,7 @@
 #include <string>
 #include <sys/types.h>
 
-#include "common.h"
+#include "getsysctl.h"
 #include "memory.h"
 #include "../luts.h"
 #include "../config.h"
